Start an OrbitalPhysics class
The core idea for this iteration is that the orbits are truly about
energy and angular momentum. Perhaps when we add in "burns", rather than
trying to iterate the actual physics of forces, we can convert the burn into
work and torque. Applying work and torque to the orbital physics bit will
update the orbital elements on the fly.

Also add in a bunch of additional unit stuff to help make this work.

Signed-off-by: Alex Shaver <alexshaver@gmail.com>