Mix of implementation - add physics cache
Physics cache was added to solve the problem of needing to
pass semiMajorAxis to period. We _could_ let that be free
and allow the user to pass in the sMA just calculated,
but that would break class invariance, I think. So the cache
will be calculated on each physics update, and then that
also reduces calculation cost by reusing as many of the
orbital params as possible.

Using this moment as a check-in to refactor the
OrbitalPhysicsTest into a matrix-based test.

Signed-off-by: Alex Shaver <alexshaver@gmail.com>