Complete Sweep and Inclination; stub Matrix Test
To fully test out the variety of orbits, we
really need to create a test matrix. But to do that
we need to finish the implementation of orbit->state
and state->orbit for enough "example" cases that
we can use the other to generate the data. e.g. use
orbital elements to generate a state vector, then
check that this math of going from state->orbit aligns
with expected elements.

So I started a Matrix Test where that stuff would go,
but generating the data by hand is not feasible.

Furthermore, the only way I can feel reasonably confident
that even this is right is to build a visualization tool,
maybe something in python, to visualize various orbits and
do a by-eye check that the behaviour looks right.

I may add another test file in that just has a few simple
orbits with very simplified numbers to do generalized logic checks
and let these files dig into the maths.

Signed-off-by: Alex Shaver <alexshaver@gmail.com>