Add longitude of asc. node
Going further requires the eccentricity vector, which, to calculate, will
require us to reuse the State Vector, which we already don't have
by this point in the tooling. However, the ecc vec is crucial
to a part I've been thinking about breaking out on its own,
which is the orbital anomaly solver.

Signed-off-by: Alex Shaver <alexshaver@gmail.com>