Create a python utility to generate data sets
In order to compare our results to "realistic" ones, we need to
generate some data driven by straightforward acceleration-based
physics. This contains a core system that will allow for
data generation. From the initial data sampled, a simple
unit-circle orbit, one can observe drift (orbit seems to drift
larger over time) in this approach, but should still be valuable
for comparing bulk behaviour against our orbital approach.

Using HDF5 file format to store data (data not stored in repo
at this time, due to file size) - this should preserve binary
fidelity of the values, plus is compatible with pandas, C++.

The next iteration will be to generate several tables in
the data file for different test scenarios, and possibly some
analysis of them. Would like to know if the drift value converges
as the timestep diminishes and what an acceptable timestep would be.

Also as a note, on my computer, generating 100 seconds of data
for the unit circle @ 1 ms resolution (20 ms recorded steps)
took about 0.8125 seconds, so data generation rates are really
fine, even if we were to go to microsecond resolution

Signed-off-by: Alex Shaver <alexshaver@gmail.com>