Analyze more physical data set
Starting with orbits that are more physical (i.e. something with a period of
O(1e4) seconds, distances of 1e6, etc.) We see a much smaller drift in radius,
worst case is like 8mm over 1Mm over 100 seconds (with approximately an error
growing cubically). Also, in terms of time to generate, O(100 sec) to generate
100 sec of data at 1µs precision is reasonable to generate 24 test cases in
40 minutes (approximately).
So now we can just move forward with generating data sets for our example
problems (which we have yet to define)

Signed-off-by: Alex Shaver <alexshaver@gmail.com>