Fully sketch out how library will operate.
The Orbit (or OrbitImpl) class will be a broker for delta V changes
as well as a source of state updates. It will hold a Kernel, an
ElementsGenerator, and one implementation of the Solvers.

Upon Delta V updates, it will update the Kernel, refresh the
cache in ElementsGenerator with the updated Kernel, and then
refresh or replace the Solver as needed. When asked for position,
it will consult the solver who should already have the
most up to date information at that time.

Signed-off-by: Alex Shaver <alexshaver@gmail.com>