Compose tests for Mean Anomaly
Learning here: it's fine to keep unit tests semi-trivial if there are better
integration tests to come. Rather than hand composing a bunch of vectors and
tediously solving the maths by hand here, have a simple and less simple case
and know that we'll have an integration test to verify that everything works
together.

Signed-off-by: Alex Shaver <alexshaver@gmail.com>