Store State at Epoch, Epoch in Kernel
Not completely sure this is where they belong, but it seems
like the right place for them. What we'll need to do is as
part of the update cycle when applying delta-V, is to make
sure that the epoch details are updated with each tick.

Also the IDE is going crazy on formatting right now and
I need to just set a consistent clang format file and move on
from there. So next commit might just be a reformat

Signed-off-by: Alex Shaver <alexshaver@gmail.com>